requirements.txt
httpx[http2]
python-dotenv